find minimum on surrogate model
Previously, the best evaluation was used as the minimum.
This is problematic due to noisyness. Additionally, the EI criterion
doesn't necessarily go for the minimum if uncertain regions are
available.

Now, the final surrogate model is used as the objective function for a
classic derivative-free optimization method, which should help get
closer to the true minimum. It cannot hurt. In principle a
derivative-based optimization method could be used as well,
but the gradients would have to first be implemented in the kernel.