use Upper Confidence Bound for suggesting optimum
In the final suggestion phase, the global minimum is searched for the
model. However, this may lead to areas hat have high variance, because
they are far from previous samples.

Most users want to tune algorithms that are mostly good most of the
time, not very good occasionally. Thus, variance needs to be penalized.

This penalization is achieved by searching the minimum upper confidence
bound. The size of the confidence band is user-configurable, but set to
a 1-sigma confidence bound by default.

The old behaviour can be achieved by setting the confidence bound to
zero, this staying at the mean.
If the objective is log-transformed, this will be the median instead.